Fix: RouteRepositoryImpl 클래스 findReservedRoutes 함수 수정 api 문서 변경에 따라 select에서 route.status 추가

+++ BE/src/main/java/com/lisade/togeduck/dto/response/RouteDetailResponse.java
@@ -16,6 +16,6 @@ public class RouteDetailResponse {
     LocalTime expectedAt;
     LocalTime arrivalAt;
     Integer totalSeats;
-    Long reservedSeats;
+    Integer reservedSeats;
     Integer cost;
 }
